+++ src/webpack/getRules.js
@@ -1,5 +1,6 @@
 import path from 'path'
 import Utils from '../utils/'
+import pxtorem from 'postcss-pxtorem'
 import lessVariableInjection from 'less-variable-injection'
 const dirSrc = path.join(process.cwd(), 'src')
 const dirNodeModule = /node_modules/
@@ -34,7 +35,7 @@ export default () => {
     let babelOptions = {
         babelrc: false,
         cwd: path.resolve(Utils.path.parentDir, 'node_modules'),
-        presets: [['@babel/preset-env'], '@babel/preset-react'],
+        presets: [['@babel/preset-env', { modules: 'commonjs' }], '@babel/preset-react'],
         plugins: [
             [
                 'babel-plugin-react-scoped-css',
@@ -147,6 +148,35 @@ export default () => {
             },
             { loader: 'scoped-css-loader' }
         ]
+    }
+
+    //开启 px 转 rem
+    if (configJson.pxtorem) {
+        console.info('启用postcss-loader')
+        const postcssLoaderConfig = {
+            loader: 'postcss-loader',
+            options: {
+                postcssOptions: {
+                    plugins: []
+                }
+            }
+        }
+        const pxtoremConfig = {
+            rootValue: 37.5,
+            propList: ['*'],
+            exclude: /node_modules/i
+        }
+
+        if (configJson.pxtorem === true) {
+            postcssLoaderConfig.options.postcssOptions.plugins.push(pxtorem(pxtoremConfig))
+        } else if (typeof configJson.pxtorem === 'number') {
+            pxtoremConfig.rootValue = configJson.pxtorem
+            postcssLoaderConfig.options.postcssOptions.plugins.push(pxtorem(pxtoremConfig))
+        } else {
+            postcssLoaderConfig.options.postcssOptions.plugins.push(pxtorem(Object.assign(pxtoremConfig, configJson.pxtorem)))
+        }
+        less.use.splice(3, 0, postcssLoaderConfig)
+        css.use.splice(3, 0, postcssLoaderConfig)
     }
 
     file = {